back/requirements.txt
attrs==21.2.0
click==8.0.1
Flask==2.0.1
Flask-PyMongo==2.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
mongoengine==0.23.1
packaging==21.0
Pillow==8.3.2
pluggy==1.0.0
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
toml==0.10.2
Werkzeug==2.0.1
